% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.dccav.r
\name{anova.dccav}
\alias{anova.dccav}
\title{Community- and Species-Level Permutation Test in Double Constrained Correspondence Analysis (dc-CA)}
\usage{
\method{anova}{dccav}(object, permutations = 999, by = NULL)
}
\arguments{
\item{object}{an object from \code{\link{dc_CA_vegan}}.}

\item{permutations}{a list of control values for the permutations for species and sites (species first, sites second,
for traits and environment) as
 returned by the function \code{\link[permute]{how}}, or
 the number of permutations required (default 999,
 or a two-vector with the number for the species-level test first and
 that for the sites-level second), or
 a list of two permutation matrices (again, species first, sites second)
 where each row gives the permuted indices.}

\item{by}{character \code{"axis"} which performs a series of test, one for each axis,
with the eigenvalue of the axis as  test statistic.
Default: \code{NULL} which set the test statistic to the inertia named \code{constraintsTE}
in the inertia element of \code{\link{dc_CA_vegan}}). The interpretation of this inerita is,
at the species-level, the environmentally constrained inertia explained by the traits (without trait covariates) and,
at the community-level, the trait-constrained inertia
explained by the environmental predictors (without covariates).
The default (\code{NULL}) is computationally quicker as it avoids computation of an svd of permuted data sets.}
}
\value{
A list of 3 of structures as from \code{\link[vegan]{anova.cca}}. The elements are \code{c("species","sites","max")}
}
\description{
\code{anova.dccav} performs the community- and species-level permutation tests of dc-CA
and combines these with the 'max test', which takes the maximum of the \emph{P}-values.
The test uses residualized response permutation for the community-level, as implemented in
\code{\link[vegan]{anova.cca}} and
residualized predictor permutation (ter Braak 2022) for the species-level, which is robust
against differences in species total abundance in the \code{response} in \code{\link{dc_CA_vegan}} (ter Braak & te Beest, 2022)
The function arguments are similar to (but more restrictive than) those of \code{\link[vegan]{anova.cca}}.
}
\details{
The community-level test is obtained by applying \code{anova} to \code{object$RDAonEnv}
 using \code{\link[vegan]{anova.cca}}.
This performs residualized response permutation but this permutation method performs about equal
to residualized predictor permutation in the equi-weight case of \code{\link{dc_CA_vegan}}.
The species-level test is obtained by applying residualized predictor permutation to
the species-niche centroids with respect to orthonormalized environmental variables using
the functions \code{\link{anova_species}} with the internal function \code{randperm_eX0sqrtw}
(taken from ter Braak 2022 with re-structured return value and used for
the residual predictor permutation (ter Braak 2022, ter
Braak & te Beest, 2022).
The \code{dcCA} package has the function \code{anova_sites}
which does residual predictor permutation, also using \code{randperm_eX0sqrtw}, but this
function is native \code{R} code and thus a factor of 20 or so slower than \code{\link[vegan]{anova.cca}}.
}
\examples{
data("dune_trait_env")
# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites
mod <- dc_CA_vegan(formulaEnv = ~A1+Moist+Mag+Use+Manure,
                   formulaTraits = ~. ,
                   response = dune_trait_env$comm[, -1],  # must delete "Sites"
                   dataEnv =dune_trait_env$envir,
# delete "Species", "Species_abbr" from traits and
# use all remaining variables due to formulaTraits = ~. (the default)
                   dataTraits =dune_trait_env$traits[,-c(1,2)],
                   verbose = TRUE)
anova(mod)

a_species <- anova_species(mod)
a_species
# anova_species can be used for manual forward selection of
# trait variables, as done for environmental variables in the demo dune_FS_dcCA.r,
# based on the first eigenvalue and its significance
# and adding the first axis of the constrained species scores from mod to the Condition
# of a new mod.
(eig1_and_pval <- c(eig = a_species$eig[1], pval = a_species$table$`Pr(>F)`[1]))
a_species$eig
#
anova_sites(mod)

mod$eigenvalues / anova_species(mod)$eig[seq_along(mod$eigenvalues)]
mod$eigenvalues / anova_sites(mod)$eig[seq_along(mod$eigenvalues)]








# Benchmark the community-level test
#microbenchmark::microbenchmark(anova(mod$RDAonEnv, by = "axis"), anova_sites(mod, by ="axis"),times = 1)
#microbenchmark::microbenchmark(anova(mod$RDAonEnv), anova_sites(mod), times = 2)
}
\references{
ter Braak, C.J.F. & te Beest, D.E. 2022. Testing environmental effects
on taxonomic composition with canonical correspondence analysis:
alternative permutation tests are not equal.
Environmental and Ecological Statistics. 29 (4), 849-868.
https://doi.org/10.1007/s10651-022-00545-4

ter Braak, C.J.F. (2022) Predictor versus response permutation
for significance testing in weighted regression and redundancy analysis.
Journal of statistical computation and simulation, 92, 2041-2059.
 https://doi.org/10.1080/00949655.2021.2019256
}
