% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_CAForwardSelectionTraits.R
\name{dc_CAForwardSelectionTraits}
\alias{dc_CAForwardSelectionTraits}
\title{Forward selection of traits using dc-CA  with a fixed environmental formula}
\usage{
dc_CAForwardSelectionTraits(
  mod,
  consider = NULL,
  permutations = 999,
  n_axes = "all",
  initial_trait_model = "1",
  test = 2,
  threshold_P = 0.1,
  PvalAdjustMethod = "holm",
  max_step = 10,
  verbose = TRUE
)
}
\arguments{
\item{mod}{initial dA-CAmodel with at least on one trait, and an 
environmental formula to be used throughout the selection.}

\item{consider}{character vector of names in \code{mod$data$dataTraits}
to consider of addition. Default \code{NULL} for all variables in 
\code{mod$data$dataTraits}.}

\item{permutations}{a list of control values for the permutations as 
returned by the function \code{\link[permute]{how}}, or the number of 
permutations required (default 999), or a permutation matrix where each 
row gives the permuted indices. 
If \code{test = 2}, a two-vector with the number for the
species-level test first and that for the sites-level second), or
a list of two permutation matrices (again, species first, sites second)
where each row gives the permuted indices.}

\item{n_axes}{number of dc-CA eigenvalues to select upon.
The sum of \code{n_axes} eigenvalues is taken as criterion.
Default \code{"full"} for selection without dimension reduction to 
\code{n_axes}.
If \code{n_axes =1}, selection is on the first eigenvalue of the dc-CA 
for selection of variables that form an optimal one-dimensional model.}

\item{initial_trait_model}{character specifying what should be inside 
\code{Condition()}. Default: \code{"1"} (nothing, the intercept only).
Examples: \code{"region"} for a within-region analysis or
\code{"A*B"} for a within analysis specified by the interaction 
of factors \code{A} and \code{B}, with \code{region, A, B} in the data.}

\item{test}{integer value for statistical testing; default: \code{2}. 
For \code{test = 0,1,2}, no test, site-based permutation test and site- 
& species-based permutation tests, respectively.
With \code{test = 1}, the test is likely liberal.}

\item{threshold_P}{significance level, after adjustment for testing 
multiplicity, for addition of a variable to the model.}

\item{PvalAdjustMethod}{method for correction for multiple testing 
in  \code{\link[stats]{p.adjust}}, default \code{"holm"}, which
is an improved version Bonferroni.}

\item{max_step}{maximal number of variables selected.}

\item{verbose}{show progress, default: \code{TRUE}.}
}
\value{
list with three elements: \code{final...} with selected variables,
\code{model_final}, and \code{process} with account of the selection process
}
\description{
Forward selection of traits using dc-CA  with a fixed environmental formula
}
\details{
The selection is on the basis of the additional fit (inertia) of a variable 
given the variables already in the model.

The \code{n_axes} argument is considered experimental.
}
\examples{
data("dune_trait_env")

# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites

mod <- dc_CA(formulaEnv = ~ A1,
             formulaTraits = ~ F + R + N + L,
             response = dune_trait_env$comm[, -1],  # must delete "Sites"
             dataEnv = dune_trait_env$envir[, -c(1, 7:10)], # without Sites, X, Y etc.
             dataTraits = dune_trait_env$traits,
             verbose = FALSE)

out1 <- dc_CAForwardSelectionEnv(mod, verbose = FALSE) 
names(out1)
out1$process
out1$finalWithOneExtra
out1$model_final

mod2 <- dc_CA(formulaEnv = ~ Moist + Mag,
              formulaTraits = ~ F,
              response = dune_trait_env$comm[, -1],  # must delete "Sites"
              dataEnv = dune_trait_env$envir,
              dataTraits = dune_trait_env$traits[, -c(1, 2)],
              verbose = FALSE)

out2 <- dc_CAForwardSelectionTraits(mod2, verbose = FALSE) 
names(out2)
out2$process
out2$finalWithOneExtra
out2$model_final
}
