% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_env_axes.r
\name{f_env_axes}
\alias{f_env_axes}
\title{Utility function: obtain column (sites) related scores}
\usage{
f_env_axes(out, which_cor = "in model")
}
\arguments{
\item{out}{object from \code{\link{dc_CA_vegan}}}

\item{which_cor}{character or names of environmental variables
for which inter-set correlations must calculated.
Default: "in_model" for all environmental variables in the model specified by \code{formulaEnv}.}
}
\description{
Utility function: obtain column (sites) related scores
}
