% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_trait_axes.r
\name{f_trait_axes}
\alias{f_trait_axes}
\title{Utility function: obtain column (species) related Scores}
\usage{
f_trait_axes(out, which_cor = "in model")
}
\arguments{
\item{out}{object from \code{\link{dc_CA_vegan}}}

\item{which_cor}{character or names of traits
for which inter-set correlations must calculated.
Default: "in_model" for all traits and variables in the model specified by \code{formulaTraits}.}
}
\description{
Utility function: obtain column (species) related Scores
}
