% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dcCA.r
\name{plot_dcCA}
\alias{plot_dcCA}
\title{Plot a single dc-CA axis with CWMs, SNCs, trait and environment scores.}
\usage{
plot_dcCA(
  object,
  axis = 1,
  gradient_description = "correlation",
  envfactor = NULL,
  traitfactor = NULL,
  nspecies = 20,
  species_groups = NULL,
  widths = c(5, 1, 1),
  newnames = NULL,
  facet = TRUE,
  remove.centroids = FALSE,
  with_lines = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{results from \code{\link{dc_CA_vegan}} of class \code{dccav}.}

\item{axis}{the axis number to get (default 1).}

\item{gradient_description}{character or 2-character vector for the trait
and environmental gradient, respectively. What to plot in the vertical line plots to describe the
dc-CA axis (trait and environmental gradients). Default: \code{correlation} for intra-set correlations
of both sets of variables with the dc-CA axis. Other values are:
c("regression","weights","tvalues", "inter_set_correlation") for regression weights, t-values and
(other) namely inter-set correlation, being the correlation of the SNCs and CWMs with the traits and
environmental variables, respectively.}

\item{envfactor}{name of row factor to display as color and lines in the CWM plot (default \code{NULL}).
The default extracts the factor from the environmental model.
If set to \code{NA}, no additional coloring and lines are displayed.}

\item{traitfactor}{name of column factor to display as color and lines in the SNC plot (default \code{NULL}).
The default extracts the factor from the trait model.
If set to \code{NA}, no additional coloring and lines are displayed.}

\item{nspecies}{integer. Default \code{20} for including a vertical species plot
with at most \code{nspecies} that have the highest contribution.}

\item{species_groups}{name of a variable in \code{dataTraits} of \code{\link{dc_CA_vegan}}. Default \code{NULL}
for no grouping. NOT yet implemented.}

\item{widths}{relative widths of the CWM-SNC plot, the correlation/weight plot
and the species plot. (see \code{\link[gridExtra]{grid.arrange}}). Default \code{c(5,1,1)}.}

\item{newnames}{a list  with two elements:
names for traits and for environmental variables, default \code{NULL} for
names derived from the result of \code{\link{scores.dccav}} with \code{tidy=TRUE}.}

\item{facet}{logical. Default \code{TRUE} for CWMs and SNCs plots in separate panels.
This parameter changes the position of the centroid names (from left to right for the environmental
centroids).
If \code{facet = TRUE} and \code{with_lines = TRUE}, the line fits ignore groups of species and of sites.}

\item{remove.centroids}{logical to remove any centroids from the plot data (default \code{FALSE}).
Can be a two-vector, \emph{e.g.} \code{c(TRUE, FALSE)} to remove only the environmental centroids.}

\item{with_lines}{logical. Default \code{FALSE}. TRUE for straight lines though groups of points.
\code{traitfactor = NA} and \code{envfactor = NA}. Centroids are not displayed in this case.}

\item{verbose}{logical. Default \code{TRUE} for plotting the result.}
}
\description{
\code{plot_dcCA} plots the CWMs and SNCs of a dc-CA axis against this axis,
 with optional centroids and colors for groups of sites and/or species if available in the data.
}
\details{
The lines with \code{with_lines=TRUE} do no use the weights in this version and
may give an extra band for an not-existing line (for missing centroids).

If you want to set new names, look at the names with all arguments default, i.e.
\code{myplot <- plot_dcCA(object)}, and then consult \code{myplot$name.list$newnames} for the
order of the names of traits and environmental variables. Note that covariates should not be in the
list of names.
Contribution (in the definition of species selection in \code{nspecies}) is defined (as in CA) as the total species abundance in the closed data
multiplied by the square of the score on the axis.

If the \code{plot_dcCA} returns the error \code{"Error in grid.Call"}, enlarge the plotting area or
use \code{verbose = FALSE} and assign the result.
}
\examples{

data("dune_trait_env")
# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites
# must delete "Sites" from response matrix or data frame
Y <- dune_trait_env$comm[,-1] # must delete "Sites"

out <- dc_CA_vegan(formulaEnv = ~A1+Moist+Use+Manure+Condition(Mag),
                   formulaTraits = ~ SLA + Height + LDMC + Seedmass +Lifespan,
                   response = Y,
                   dataEnv =dune_trait_env$envir,
                   dataTraits =dune_trait_env$traits,
                   verbose = FALSE)
plot_dcCA(out, verbose = FALSE)

}
