% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wrda.R
\name{print.wrda}
\alias{print.wrda}
\title{Print a summary of a wrda or cca0 object}
\usage{
\method{print}{wrda}(x, ...)
}
\arguments{
\item{x}{an object from \code{\link{wrda}} or \code{\link{cca0}}}

\item{...}{Other arguments passed to the function (currently ignored).}
}
\value{
No return value, results are printed to console.
}
\description{
Print a summary of a wrda or cca0 object
}
\examples{
data("dune_trait_env")

# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites
response <- dune_trait_env$comm[, -1]  # must delete "Sites"

w <- rep(1, 20) 
w[1:10] <- 8 
w[17:20] <- 0.5

object <- wrda(formula = response ~ A1 + Moist + Mag + Use + Condition(Manure),
               data = dune_trait_env$envir, 
               weights = w)
object # Proportions equal to those Canoco 5.15

mod_scores <- scores(object, display = "all")
scores(object, which_cor = c("A1", "X_lot"), display = "cor")
anova(object)
}
