% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.dcca.R
\name{FS.dcca}
\alias{FS.dcca}
\title{Forward selection of traits or environmental variables using dc-CA.}
\usage{
\method{FS}{dcca}(
  mod,
  ...,
  select = c("traits", "env"),
  consider,
  permutations = 999,
  n_axes = "all",
  initial_model = "1",
  factor2categories = TRUE,
  test = TRUE,
  threshold_P = 0.1,
  PvalAdjustMethod = "holm",
  max_step = 10,
  verbose = FALSE
)
}
\arguments{
\item{mod}{dc-CA model.}

\item{...}{unused.}

\item{select}{Character, the default \code{"traits"} selects
trait variables based on the environmental model of \code{mod}.
Set \code{select = "e"} to select environmental variables based 
on the trait model of \code{mod}.}

\item{consider}{character vector of names considered for addition, either 
trait names in \code{mod$data$dataTraits} or
environmental variable names in in \code{mod$data$dataEnv}.
The names may include transformations of predictor variables,
such as \code{log(.)},
if \code{consider} does not include factors or if \code{factor2categories=FALSE}.
If \code{consider} includes factors, such transformations
give in a error in the default setting (\code{factor2categories=TRUE}).}

\item{permutations}{a list of control values for the permutations as
returned by the function \code{\link[permute]{how}}, or the number of 
permutations required (default 999), or a permutation matrix where each row
gives the permuted indices.}

\item{n_axes}{number of eigenvalues to select upon.
The sum of \code{n_axes} eigenvalues is taken as criterion.
Default \code{"full"} for selection without dimension reduction to 
\code{n_axes}.
If \code{n_axes =1}, selection is on the first eigenvalue 
for selection of variables that form an optimal one-dimensional model.}

\item{initial_model}{character specifying what should be inside 
\code{Condition()}. Default: \code{"1"} (nothing, the intercept only).
Examples: With selection of environmental varialbles, 
\code{"region"} for a within-region analysis or
\code{"A*B"} for a within analysis specified by the interaction 
of factors \code{A} and \code{B}, with \code{region, A, B} in the data.
Analogously, a trait or trait factor can be the initial model.}

\item{factor2categories}{logical, default \code{TRUE}, to convert
factors to their categories (set(s) of indicator values). 
If \code{FALSE}, the selection uses, the fit of a factor
divided by its number of categories minus 1.}

\item{test}{logical; default: \code{TRUE}.}

\item{threshold_P}{significance level, after adjustment for testing 
multiplicity, for addition of a variable to the model.}

\item{PvalAdjustMethod}{method for correction for multiple testing 
in  \code{\link[stats]{p.adjust}}, default \code{"holm"}, which
is an improved version Bonferroni.}

\item{max_step}{maximal number of variables selected.}

\item{verbose}{show progress, default: \code{TRUE}.}
}
\value{
list with three elements: \code{final...} with selected variables,
\code{model_final}, and \code{process} with account of the selection process
}
\description{
Forward selection of traits or environmental variables using dc-CA.
}
\details{
The selection is on the basis of the additional fit (inertia) of a variable 
given the variables already in the model.

The function \code{FS} does implement the max test and may thus 
be liberal. It is recommended to test the final model (element
\code{model_ final} of the returned object) by \code{anova}.
}
\examples{
data("dune_trait_env")

# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites

mod <- dc_CA(formulaEnv = ~ A1,
             formulaTraits = ~ F + R + N + L,
             response = dune_trait_env$comm[, -1],  # must delete "Sites"
             dataEnv = dune_trait_env$envir[, -c(1, 7:10)], # without Sites, X, Y etc.
             dataTraits = dune_trait_env$traits,
             verbose = FALSE)

out1 <- dc_CAForwardSelectionEnv(mod, verbose = FALSE) 
names(out1)
out1$process
out1$finalWithOneExtra
out1$model_final

mod2 <- dc_CA(formulaEnv = ~ Moist + Mag,
              formulaTraits = ~ F,
              response = dune_trait_env$comm[, -1],  # must delete "Sites"
              dataEnv = dune_trait_env$envir,
              dataTraits = dune_trait_env$traits[, -c(1, 2)],
              verbose = FALSE)

out2 <- dc_CAForwardSelectionTraits(mod2, verbose = FALSE) 
names(out2)
out2$process
out2$finalWithOneExtra
out2$model_final
}
