% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.wrda.R
\name{FS}
\alias{FS}
\title{Default forward selection function.}
\usage{
FS(mod, ...)
}
\arguments{
\item{mod}{A fitted model.}

\item{...}{Further arguments passed to other methods.}
}
\value{
The results from applying forward selection on the fitted model.
}
\description{
Default forward selection function.
}
