% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dcca.r
\name{predict.dcca}
\alias{predict.dcca}
\title{Prediction for double-constrained correspondence analysis (dc-CA)}
\usage{
\method{predict}{dcca}(
  object,
  type = c("env", "traits", "response", "reg_env", "reg_traits"),
  rank = "full",
  newdata = NULL
)
}
\arguments{
\item{object}{return value of \code{\link{dc_CA}}.}

\item{type}{type of prediction, c("env", "traits","response","reg_env","reg_trait") for
environmental values, values of traits,
regression coefficients for environmenal and trait predictors.}

\item{rank}{rank or number of axes to use. Default "full" for all axes (no rank-reduction).}

\item{newdata}{An obtional data frame in which to look for variable with which to predict. If omitted,
the fitted values are generated.}
}
\description{
Prediction of traits from environment and environment from traits

With \code{type = "traits"} and \code{newdata=NULL}, predict gives the fitted mean traits,
\emph{i.e.} the fitted community weighted means.
With \code{type = "env"} and \code{newdata=NULL}, predict gives the fitted mean environment,
\emph{i.e.} the fitted species niche centroids.
}
